backend/requirements.txt
django<3
djangorestframework
djangorestframework-camel-case
django-extra-fields
django-cors-headers
drf_yasg
Pillow
django-jet
python_http_client
drf-jwt
gunicorn